user generated types
