--- src/graphql/call.ts
@@ -1,124 +1,0 @@
-import { check, fail, group, sleep } from 'k6';
-import exec from 'k6/execution';
-
-import { User } from './support/user';
-import { getAsyncClientApi } from '../support/graphql';
-import { randomIntBetween } from '../utils/helperFunctions';
-import { GenericQueryResponse, SharedData } from '../utils/sharedType';
-export async function callTest(sharedData: SharedData) {
-  if (!sharedData.users) {
-    fail(`User not set`);
-  }
-  if (!sharedData.testCall) {
-    fail(`Test call not set`);
-  }
-  const testCall = sharedData.testCall;
-  const apiAsyncClient = getAsyncClientApi(sharedData.graphqlUrl);
-  const user = new User(apiAsyncClient);
-  sleep(randomIntBetween(5, 20));
-  const currentUser = sharedData.users[exec.vu.iterationInScenario];
-  const userToken = await user.getUserToken(`${currentUser.sessionId}`);
-  group('Call Test', () => {
-    group('Calls query should return active calls', async () => {
-      const response = await apiAsyncClient(
-        JSON.stringify({
-          query: `
-          query Calls($filter: CallsFilter) {
-            calls(filter: $filter) {
-              id
-              title
-              shortCode
-            }
-          }`,
-          variables: {
-            filter: {
-              isActive: true,
-              isEnded: false,
-            },
-          },
-        }),
-        userToken
-      );
-
-      check(response, {
-        'Calls query status is 200': (res) => res.status === 200,
-        'Calls query returned active calls': (res) => {
-          try {
-            const data = res.json() as GenericQueryResponse;
-
-            return data.data?.calls.length > 0 && !!data.data?.calls[0].id;
-          } catch (error) {
-            fail(`SCENARIO: ${exec.scenario.name} Executing callTest calls query VU_ID: ${exec.vu.idInTest}
-          Error response callTest calls query ${response.status} ${response?.body} ${response?.error} ${response?.error_code} ${error}`);
-          }
-        },
-      });
-    });
-
-    group('Call query should return call details', async () => {
-      const response = await apiAsyncClient(
-        JSON.stringify({
-          query: `
-          query Call($callId: Int!) {
-            call(callId: $callId) {
-              id
-              title
-              shortCode
-            }
-          }`,
-          variables: {
-            callId: testCall.id,
-          },
-        }),
-        userToken
-      );
-
-      check(response, {
-        'Call query status is 200': (res) => res.status === 200,
-        'Call query return call details': (res) => {
-          try {
-            const data = res.json() as GenericQueryResponse;
-
-            return !!data.data?.call.id;
-          } catch (error) {
-            fail(`SCENARIO: ${exec.scenario.name} Executing callTest call query VU_ID: ${exec.vu.idInTest}
-          Error response callTest call query ${response.status} ${response?.body} ${response?.error} ${response?.error_code} ${error}`);
-          }
-        },
-      });
-    });
-
-    group('Template query should return template details', async () => {
-      const response = await apiAsyncClient(
-        JSON.stringify({
-          query: `
-          query Template($templateId: Int!) {
-            template(templateId: $templateId) {
-              templateId
-              description
-              name
-            }
-          }`,
-          variables: {
-            templateId: testCall.templateId,
-          },
-        }),
-        userToken
-      );
-
-      check(response, {
-        'Template query status is 200': (res) => res.status === 200,
-        'Template query return template details': (res) => {
-          try {
-            const data = res.json() as GenericQueryResponse;
-
-            return !!data.data?.template.templateId;
-          } catch (error) {
-            fail(`SCENARIO: ${exec.scenario.name} Executing callTest template query VU_ID: ${exec.vu.idInTest}
-          Error response callTest template query ${response.status} ${response?.body} ${response?.error} ${response?.error_code} ${error}`);
-          }
-        },
-      });
-    });
-  });
-}
