+++ src/graphql/gql/answerTopic.graphql
@@ -1,0 +1,17 @@
+mutation AnswerTopic(
+  $questionaryId: Int!
+  $topicId: Int!
+  $answers: [AnswerInput!]!
+  $isPartialSave: Boolean
+) {
+  answerTopic(
+    questionaryId: $questionaryId
+    topicId: $topicId
+    answers: $answers
+    isPartialSave: $isPartialSave
+  ) {
+    questionId
+    answer
+    answerId
+  }
+}
