+++ src/graphql/gql/getCall.graphql
@@ -1,0 +1,7 @@
+query Call($callId: Int!) {
+  call(callId: $callId) {
+    id
+    title
+    shortCode
+  }
+}
