+++ src/graphql/gql/getMe.graphql
@@ -1,0 +1,11 @@
+query Me {
+  me {
+    id
+    created
+    roles {
+      id
+      shortCode
+      title
+    }
+  }
+}
