+++ src/graphql/gql/getProposalById.graphql
@@ -1,0 +1,7 @@
+query ProposalById($proposalId: String!) {
+  proposalById(proposalId: $proposalId) {
+    proposalId
+    primaryKey
+    callId
+  }
+}
