+++ src/graphql/gql/getProposalStatus.graphql
@@ -1,0 +1,8 @@
+query ProposalStatus($proposalStatusId: Int!) {
+  proposalStatus(proposalStatusId: $proposalStatusId) {
+    id
+    isDefault
+    name
+    shortCode
+  }
+}
