+++ src/graphql/gql/getProposals.graphql
@@ -1,0 +1,13 @@
+query getProposalsMinimal($filter: ProposalsFilter) {
+  proposals(filter: $filter) {
+    proposals {
+      primaryKey
+      proposalId
+      title
+      submitted
+      status {
+        name
+      }
+    }
+  }
+}
