+++ src/graphql/gql/getSettings.graphql
@@ -1,0 +1,6 @@
+query Settings {
+  settings {
+    description
+    id
+  }
+}
